requirements.txt
# ── Computer Vision ───────────────────────────────────────────────────────────
opencv-python>=4.8.0
mediapipe>=0.10.0

# ── Deep Learning ─────────────────────────────────────────────────────────────
tensorflow>=2.12.0

# ── Machine Learning ──────────────────────────────────────────────────────────
scikit-learn>=1.3.0

# ── Numerical ─────────────────────────────────────────────────────────────────
numpy>=1.24.0

# ── Image / Font Rendering ────────────────────────────────────────────────────
Pillow>=10.0.0

# ── Translation & Text-to-Speech ─────────────────────────────────────────────
# NOTE: Use exactly 4.0.0-rc1 — newer versions are broken
googletrans==4.0.0-rc1
gTTS>=2.3.2

# ── Audio Playback ────────────────────────────────────────────────────────────
pygame>=2.5.0

# ── Visualization (training plots) ───────────────────────────────────────────
matplotlib>=3.7.0
